+++ tests/custom-coder/tests/custom-coder.ts
@@ -1,6 +1,6 @@
 import * as anchor from "@project-serum/anchor";
 import { Spl } from "@project-serum/anchor";
-import * as assert from "assert";
+import { assert } from "chai";
 import BN from "bn.js";
 import { Keypair, SYSVAR_RENT_PUBKEY } from "@solana/web3.js";
 
@@ -34,13 +34,13 @@ describe("custom-coder", () => {
       }
     );
     const mintAccount = await program.account.mint.fetch(mintKeypair.publicKey);
-    assert.ok(
+    assert.isTrue(
       mintAccount.mintAuthority.equals(program.provider.wallet.publicKey)
     );
-    assert.ok(mintAccount.freezeAuthority === null);
-    assert.ok(mintAccount.decimals === 6);
-    assert.ok(mintAccount.isInitialized);
-    assert.ok(mintAccount.supply.toNumber() === 0);
+    assert.isNull(mintAccount.freezeAuthority);
+    assert.strictEqual(mintAccount.decimals, 6);
+    assert.isTrue(mintAccount.isInitialized);
+    assert.strictEqual(mintAccount.supply.toNumber(), 0);
   });
 
   it("Creates a token account for alice", async () => {
@@ -59,14 +59,14 @@ describe("custom-coder", () => {
     const token = await program.account.token.fetch(
       aliceTokenKeypair.publicKey
     );
-    assert.ok(token.authority.equals(program.provider.wallet.publicKey));
-    assert.ok(token.mint.equals(mintKeypair.publicKey));
-    assert.ok(token.amount.toNumber() === 0);
-    assert.ok(token.delegate === null);
-    assert.ok(token.state === 1);
-    assert.ok(token.isNative === null);
-    assert.ok(token.delegatedAmount.toNumber() === 0);
-    assert.ok(token.closeAuthority === null);
+    assert.isTrue(token.authority.equals(program.provider.wallet.publicKey));
+    assert.isTrue(token.mint.equals(mintKeypair.publicKey));
+    assert.strictEqual(token.amount.toNumber(), 0);
+    assert.isNull(token.delegate);
+    assert.strictEqual(token.state, 1);
+    assert.isNull(token.isNative);
+    assert.strictEqual(token.delegatedAmount.toNumber(), 0);
+    assert.isNull(token.closeAuthority);
   });
 
   it("Mints a token to alice", async () => {
@@ -82,8 +82,8 @@ describe("custom-coder", () => {
       aliceTokenKeypair.publicKey
     );
     const mint = await program.account.mint.fetch(mintKeypair.publicKey);
-    assert.ok(token.amount.toNumber() === 2);
-    assert.ok(mint.supply.toNumber() === 2);
+    assert.strictEqual(token.amount.toNumber(), 2);
+    assert.strictEqual(mint.supply.toNumber(), 2);
   });
 
   it("Creates a token for bob", async () => {
@@ -115,8 +115,8 @@ describe("custom-coder", () => {
     const bobToken = await program.account.token.fetch(
       bobTokenKeypair.publicKey
     );
-    assert.ok(aliceToken.amount.toNumber() === 1);
-    assert.ok(bobToken.amount.toNumber() === 1);
+    assert.strictEqual(aliceToken.amount.toNumber(), 1);
+    assert.strictEqual(bobToken.amount.toNumber(), 1);
   });
 
   it("Alice burns a token", async () => {
@@ -131,7 +131,7 @@ describe("custom-coder", () => {
       aliceTokenKeypair.publicKey
     );
     const mint = await program.account.mint.fetch(mintKeypair.publicKey);
-    assert.ok(aliceToken.amount.toNumber() === 0);
-    assert.ok(mint.supply.toNumber() === 1);
+    assert.strictEqual(aliceToken.amount.toNumber(), 0);
+    assert.strictEqual(mint.supply.toNumber(), 1);
   });
 });