+++ ts/src/idl.ts
@@ -7,7 +7,7 @@ export type Idl = {
   name: string;
   instructions: IdlInstruction[];
   state?: IdlState;
-  accounts?: IdlTypeDef[];
+  accounts?: IdlAccountDef[];
   types?: IdlTypeDef[];
   events?: IdlEvent[];
   errors?: IdlErrorCode[];
@@ -38,6 +38,7 @@ export type IdlInstruction = {
   name: string;
   accounts: IdlAccountItem[];
   args: IdlField[];
+  returns?: IdlType;
 };
 
 export type IdlState = {
@@ -77,6 +78,11 @@ export type IdlField = {
 export type IdlTypeDef = {
   name: string;
   type: IdlTypeDefTy;
+};
+
+export type IdlAccountDef = {
+  name: string;
+  type: IdlTypeDefTyStruct;
 };
 
 export type IdlTypeDefTyStruct = {